Version 0.5
Added the Save function, the playerData.txt file to hold saved values.

Save Function:

The function named setSavedInfo gets current specified data and puts it in an array. The using the saveStrings it gets saved in the playerData file

the function named getSavedInfo gets the text in the file using loadStrings and assigns them into a array.